+++ stencil-workspace/src/components/modus-content-tree/modus-tree-view-item/modus-tree-view-item.tsx
@@ -56,6 +56,9 @@ export class ModusTreeViewItem {
   /** (required) Unique tree item identifier */
   @Prop({ reflect: true }) nodeId!: string;
 
+  /** (optional) Tab Index for the tree item */
+  @Prop({ mutable: true }) tabIndexValue: string | number = 0;
+
   /**
    * @internal
    */
@@ -294,6 +297,7 @@ export class ModusTreeViewItem {
     this.options = { ...newValue };
     this.handleTreeSlotChange();
     this.updateComponent();
+    this.tabIndexValue = this.options.disableTabbing ? -1 : this.tabIndexValue;
   }
 
   rootOptions() {
@@ -382,7 +386,7 @@ export class ModusTreeViewItem {
       role: 'treeitem',
     };
     const sizeClass = `${this.classBySize.get(size || 'standard')}`;
-    const tabIndex = isDisabled ? -1 : 0;
+    const tabIndex: string | number = isDisabled ? -1 : this.tabIndexValue;
     const treeItemClass = `tree-item ${
       selected ? 'selected' : ''
     } ${sizeClass} ${isDisabled ? 'disabled' : ''} `;
@@ -422,7 +426,7 @@ export class ModusTreeViewItem {
             onKeyDown={(e) =>
               this.handleDefaultKeyDown(e, () => this.handleExpandToggle(e))
             }
-            tabIndex={expandable ? tabIndex : -1}>
+            tabindex={expandable ? tabIndex : -1}>
             <this.CustomSlot
               className="inline-flex rotate-right"
               defaultContent={<IconMap icon="chevron-down-thick" size="24" />}
@@ -448,7 +452,8 @@ export class ModusTreeViewItem {
                   this.handleDefaultKeyDown(e, () =>
                     this.handleCheckboxClick(e)
                   )
-                }></modus-checkbox>
+                }
+                tabIndexValue={tabIndex}></modus-checkbox>
             </div>
           )}
 
@@ -466,7 +471,7 @@ export class ModusTreeViewItem {
                   <modus-text-input
                     size={size == 'large' ? 'large' : 'medium'}
                     autoFocusInput={true}
-                    tabIndex={0}
+                    tabindex={0}
                     ref={(el) => (this.refLabelInput = el)}
                     value={this.label}
                     onClick={(e) => this.handleLabelInputClick(e)}