+++ CHANGELOG.md
@@ -13,6 +13,7 @@ View all releases at: <https://github.com/trimble-oss/modus-web-components/relea
 ### Added
 
 - Added format option to the Modus Date Input component.
+- Added `disableTabbing` prop to Modus Tree View and `itemTabIndex` to Modus Tree View Item
 
 ## 0.1.36 - 2023-03-15